+++ addon/utils/css-mappings/eui-popover-title.ts
@@ -1,0 +1,17 @@
+export const baseClass: string = 'euiPopoverTitle';
+
+export const paddingMapping = {
+  none: `${baseClass}--paddingNone`,
+  s: `${baseClass}--paddingSmall`,
+  m: `${baseClass}--paddingMedium`,
+  l: `${baseClass}--paddingLarge`,
+};
+
+const mapping: ComponentMapping = {
+  base: baseClass,
+  properties: {
+    paddingSize: paddingMapping,
+  },
+};
+
+export default mapping;