+++ app/modifiers/outside-click-detector.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-eui/modifiers/outside-click-detector';